django-ninja/requirements.txt
# ./requirements.txt

django-ninja==1.1.0
# django-extensions==3.2.3
mysqlclient==2.0.3